+++ GameFramework/Resource/ResourceManager.ResourceLoader.cs
@@ -633,7 +633,7 @@ namespace GameFramework.Resource
             /// <param name="offset">要加载片段的偏移。</param>
             /// <param name="length">要加载片段的长度。</param>
             /// <returns>存储加载二进制资源片段内容的二进制流。</returns>
-            public byte[] LoadBinarySegmentFromFileSystem(string binaryAssetName, long offset, int length)
+            public byte[] LoadBinarySegmentFromFileSystem(string binaryAssetName, int offset, int length)
             {
                 ResourceInfo resourceInfo = null;
                 string[] dependencyAssetNames = null;
@@ -672,7 +672,7 @@ namespace GameFramework.Resource
             /// <param name="startIndex">存储加载二进制资源片段内容的二进制流的起始位置。</param>
             /// <param name="length">要加载片段的长度。</param>
             /// <returns>实际加载了多少字节。</returns>
-            public int LoadBinarySegmentFromFileSystem(string binaryAssetName, long offset, byte[] buffer, int startIndex, int length)
+            public int LoadBinarySegmentFromFileSystem(string binaryAssetName, int offset, byte[] buffer, int startIndex, int length)
             {
                 ResourceInfo resourceInfo = null;
                 string[] dependencyAssetNames = null;
