+++ GameFramework/Resource/ResourceManager.cs
@@ -1738,7 +1738,7 @@ namespace GameFramework.Resource
                 throw new GameFrameworkException("Binary asset name is invalid.");
             }
 
-            return m_ResourceLoader.LoadBinarySegmentFromFileSystem(binaryAssetName, 0L, length);
+            return m_ResourceLoader.LoadBinarySegmentFromFileSystem(binaryAssetName, 0, length);
         }
 
         /// <summary>
@@ -1748,7 +1748,7 @@ namespace GameFramework.Resource
         /// <param name="offset">要加载片段的偏移。</param>
         /// <param name="length">要加载片段的长度。</param>
         /// <returns>存储加载二进制资源片段内容的二进制流。</returns>
-        public byte[] LoadBinarySegmentFromFileSystem(string binaryAssetName, long offset, int length)
+        public byte[] LoadBinarySegmentFromFileSystem(string binaryAssetName, int offset, int length)
         {
             if (string.IsNullOrEmpty(binaryAssetName))
             {
@@ -1777,7 +1777,7 @@ namespace GameFramework.Resource
                 throw new GameFrameworkException("Buffer is invalid.");
             }
 
-            return m_ResourceLoader.LoadBinarySegmentFromFileSystem(binaryAssetName, 0L, buffer, 0, length);
+            return m_ResourceLoader.LoadBinarySegmentFromFileSystem(binaryAssetName, 0, buffer, 0, length);
         }
 
         /// <summary>
@@ -1800,7 +1800,7 @@ namespace GameFramework.Resource
                 throw new GameFrameworkException("Buffer is invalid.");
             }
 
-            return m_ResourceLoader.LoadBinarySegmentFromFileSystem(binaryAssetName, 0L, buffer, startIndex, length);
+            return m_ResourceLoader.LoadBinarySegmentFromFileSystem(binaryAssetName, 0, buffer, startIndex, length);
         }
 
         /// <summary>
@@ -1811,7 +1811,7 @@ namespace GameFramework.Resource
         /// <param name="buffer">存储加载二进制资源片段内容的二进制流。</param>
         /// <param name="length">要加载片段的长度。</param>
         /// <returns>实际加载了多少字节。</returns>
-        public int LoadBinarySegmentFromFileSystem(string binaryAssetName, long offset, byte[] buffer, int length)
+        public int LoadBinarySegmentFromFileSystem(string binaryAssetName, int offset, byte[] buffer, int length)
         {
             if (string.IsNullOrEmpty(binaryAssetName))
             {
@@ -1835,7 +1835,7 @@ namespace GameFramework.Resource
         /// <param name="startIndex">存储加载二进制资源片段内容的二进制流的起始位置。</param>
         /// <param name="length">要加载片段的长度。</param>
         /// <returns>实际加载了多少字节。</returns>
-        public int LoadBinarySegmentFromFileSystem(string binaryAssetName, long offset, byte[] buffer, int startIndex, int length)
+        public int LoadBinarySegmentFromFileSystem(string binaryAssetName, int offset, byte[] buffer, int startIndex, int length)
         {
             if (string.IsNullOrEmpty(binaryAssetName))
             {
